add social category over "links"
